Complete rewrite (#804)
This PR collects the different components of changing GETTSIM's internal
DAG from a flat namespace to a nested one. It also updates the entire interface,
the core components that have nothing to do with the specifics of the German
taxes and transfers system have been extracted to https://github.com/ttsim-dev/ttsim.

This is a very large change spanning more than 100 individual PRs. For a near-complete 
list, see the changelog, here are some excerpts:

- [x] Infrastructure #780 
- [x] Use qualified names for rounding keys #824 
- [x] Renaming #805
- [x] Fixing the check regarding overlapping periods of policy functions
#830
- [x] Getting rid of TYPES_INPUT_VARIABLES and other cruft #833 
- [x] Use relative paths wherever possible #957 
- [x] First pass restructuring directories to separate ttsim and gettsim
(#850)
- [x] #908 
- [x] #853
- [x] #918, #883
- [x] #920
- [x] Required interface changes (parts of #781)
- [x] Convert test files from qualified name dicts to nested structure
dicts #884
- [x] Make ttsim a separate package.

---------

Co-authored-by: Lars Reimann <mail@larsreimann.com>
Co-authored-by: Marvin Immesberger <immesberger@posteo.de>
Co-authored-by: Tim Mensinger <mensingertim@gmail.com>
Co-authored-by: Felix Schmitz <33942086+felixschmitz@users.noreply.github.com>
Co-authored-by: Max Jahn <max.jahn45@gmail.com>
Co-authored-by: JuergenWiemers <juergen.wiemers@gmail.com>